requirements.txt
pandas
pyarrow
requests
python-dotenv
psycopg2-binary
